working towards #536 by moving routines into flow, so can use a flowcb entry point sanely.
