+++ sarracenia/__init__.py
@@ -561,134 +561,6 @@ class Message(dict):
         msg['_deleteOnPost'] |= set(['report'])
         msg['report'] = {'code': code, 'message': text}
 
-    # ==============================================
-    # how will the download file land on this server
-    # with all options, this is really tricky
-    # ==============================================
-
-    def updateFieldsAccepted(msg, options, urlstr, pattern, maskDir,
-                             maskFileOption, mirror, strip, pstrip, flatten):
-        """
-           Set new message fields according to values when the message is accepted.
-           
-           * options: the sarracenia.config instance giving all the current configuration.
-           * urlstr: the urlstr being matched (baseUrl+relPath+sundew_extension)
-           * pattern: the regex that was matched.
-           * maskDir: the current directory to base the relPath from.
-           * maskFileOption: filename option value (sundew compatibility options.)
-           * strip: number of path entries to strip from the left side of the path.
-           * pstrip: pattern strip regexp to apply instead of a count.
-           * flatten: a character to replace path separators with toe change a multi-directory 
-             deep file name into a single long file name
-        """
-
-        # relative path by default mirror
-        if type(maskDir) is str:
-            # trying to subtract maskDir if present in relPath...
-            # occurs in polls a lot.
-            if maskDir in msg['relPath']:
-                relPath = '%s' % msg['relPath'].replace(maskDir, '', 1)
-
-            # sometimes the same, just the leading / is missing.
-            elif maskDir[1:] in msg['relPath']:
-                relPath = '%s' % msg['relPath'].replace(maskDir[1:], '', 1)
-            else:
-                relPath = '%s' % msg['relPath']
-        else:
-            relPath = '%s' % msg['relPath']
-
-        if options.baseUrl_relPath:
-            u = urllib.parse.urlparse(msg['baseUrl'])
-            relPath = u.path[1:] + '/' + relPath
-
-        # FIXME... why the % ? why not just assign it to copy the value?
-        if 'rename' in msg: relPath = '%s' % msg['rename']
-
-        token = relPath.split('/')
-        filename = token[-1]
-
-        # if provided, strip (integer) ... strip N heading directories
-        #         or  pstrip (pattern str) strip regexp pattern from relPath
-        # cannot have both (see setting of option strip in sr_config)
-
-        if strip > 0:
-            try:
-                token = token[strip:]
-
-            # strip too much... keep the filename
-            except:
-                token = [filename]
-
-        # strip using a pattern
-
-        elif pstrip != None:
-
-            #MG FIXME Peter's wish to have replacement in pstrip (ex.:${SOURCE}...)
-            try:
-                relstrip = re.sub(pstrip, '', relPath, 1)
-            except:
-                relstrip = relPath
-
-            # if filename dissappear... same as numeric strip, keep the filename
-            if not filename in relstrip: relstrip = filename
-            token = relstrip.split('/')
-
-        # if flatten... we flatten relative path
-        # strip taken into account
-
-        if flatten != '/':
-            filename = flatten.join(token)
-            token[-1] = [filename]
-
-        if maskFileOption is not None:
-            try:
-                filename = options.sundew_getDestInfos(msg, maskFileOption,
-                                                       filename)
-            except:
-                logger.error("problem with accept file option %s" %
-                             maskFileOption)
-            token[-1] = [filename]
-
-        # not mirroring
-
-        if not mirror:
-            token = [filename]
-
-        # uses current dir
-
-        #if options.currentDir : new_dir = options.currentDir
-        if maskDir:
-            new_dir = options.set_dir_pattern(maskDir, msg)
-        else:
-            new_dir = ''
-
-        if options.baseDir:
-            if new_dir:
-                d = new_dir
-            elif options.post_baseDir:
-                d = options.set_dir_pattern(options.post_baseDir, msg)
-            else:
-                d = None
-
-            if d:
-                for f in ['link', 'oldname', 'newname']:
-                    if f in msg:
-                        msg[f] = msg[f].replace(options.baseDir, d, 1)
-
-        # add relPath
-
-        if len(token) > 1:
-            new_dir = new_dir + '/' + '/'.join(token[:-1])
-
-        new_dir = options.set_dir_pattern(new_dir, msg)
-        # resolution of sundew's dirPattern
-
-        tfname = filename
-        # when sr_sender did not derived from sr_subscribe it was always called
-        new_dir = options.sundew_dirPattern(pattern, urlstr, tfname, new_dir)
-
-        msg.updatePaths(options, new_dir, filename)
-
     def updatePaths(msg, options, new_dir, new_file):
         """
         set the new\_ fields in the message based on changed file placement.
