+++ sarracenia/config.py
@@ -1576,106 +1576,6 @@ class Config:
 
         return ndestDir
 
-    def sundew_getDestInfos(self, msg, currentFileOption, filename):
-        """
-        modified from sundew client
-
-        WHATFN         -- First part (':') of filename 
-        HEADFN         -- Use first 2 fields of filename
-        NONE           -- Use the entire filename
-        TIME or TIME:  -- TIME stamp appended
-        DESTFN=fname   -- Change the filename to fname
-
-        ex: mask[2] = 'NONE:TIME'
-        """
-        if currentFileOption == None: return filename
-
-        timeSuffix = ''
-        satnet = ''
-        parts = filename.split(':')
-        firstPart = parts[0]
-
-        if 'sundew_extension' in msg.keys():
-            parts = [parts[0]] + msg['sundew_extension'].split(':')
-            filename = ':'.join(parts)
-
-        destFileName = filename
-
-        for spec in currentFileOption.split(':'):
-            if spec == 'WHATFN':
-                destFileName = firstPart
-            elif spec == 'HEADFN':
-                headParts = firstPart.split('_')
-                if len(headParts) >= 2:
-                    destFileName = headParts[0] + '_' + headParts[1]
-                else:
-                    destFileName = headParts[0]
-            elif spec == 'SENDER' and 'SENDER=' in filename:
-                i = filename.find('SENDER=')
-                if i >= 0: destFileName = filename[i + 7:].split(':')[0]
-                if destFileName[-1] == ':': destFileName = destFileName[:-1]
-            elif spec == 'NONE':
-                if 'SENDER=' in filename:
-                    i = filename.find('SENDER=')
-                    destFileName = filename[:i]
-                else:
-                    if len(parts) >= 6:
-                        # PX default behavior : keep 6 first fields
-                        destFileName = ':'.join(parts[:6])
-                        #  PDS default behavior  keep 5 first fields
-                        if len(parts[4]) != 1:
-                            destFileName = ':'.join(parts[:5])
-                # extra trailing : removed if present
-                if destFileName[-1] == ':': destFileName = destFileName[:-1]
-            elif spec == 'NONESENDER':
-                if 'SENDER=' in filename:
-                    i = filename.find('SENDER=')
-                    j = filename.find(':', i)
-                    destFileName = filename[:i + j]
-                else:
-                    if len(parts) >= 6:
-                        # PX default behavior : keep 6 first fields
-                        destFileName = ':'.join(parts[:6])
-                        #  PDS default behavior  keep 5 first fields
-                        if len(parts[4]) != 1:
-                            destFileName = ':'.join(parts[:5])
-                # extra trailing : removed if present
-                if destFileName[-1] == ':': destFileName = destFileName[:-1]
-            elif re.compile('SATNET=.*').match(spec):
-                satnet = ':' + spec
-            elif re.compile('DESTFN=.*').match(spec):
-                destFileName = spec[7:]
-            elif re.compile('DESTFNSCRIPT=.*').match(spec):
-                old_destfn_script = self.destfn_script
-                saved_new_file = msg['new_file']
-                msg['new_file'] = destFileName
-                self.destfn_script = None
-                script = spec[13:]
-                self.execfile('destfn_script', script)
-                if self.destfn_script != None:
-                    ok = self.destfn_script(self)
-                destFileName = msg['new_file']
-                self.destfn_script = old_destfn_script
-                msg['new_file'] = saved_new_file
-                if destFileName == None: destFileName = old_destFileName
-            elif spec == 'TIME':
-                timeSuffix = ':' + time.strftime("%Y%m%d%H%M%S", time.gmtime())
-                if 'pubTime' in msg:
-                    timeSuffix = ":" + msg['pubTime'].split('.')[0]
-                if 'pubTime' in msg:
-                    timeSuffix = ":" + msg['pubtime'].split('.')[0]
-                    timeSuffix = timeSuffix.replace('T', '')
-                # check for PX or PDS behavior ...
-                # if file already had a time extension keep his...
-                if len(parts[-1]) == 14 and parts[-1][0] == '2':
-                    timeSuffix = ':' + parts[-1]
-
-            else:
-                logger.error("Don't understand this DESTFN parameter: %s" %
-                             spec)
-                return (None, None)
-        return destFileName + satnet + timeSuffix
-
     # modified from metpx SenderFTP
     def sundew_matchPattern(self, BN, EN, BP, keywd, defval):
